+++ applications/foobar/status.sh
@@ -10,4 +10,6 @@ check_env_vars
 # Log the action to Google Cloud Logging
 log_to_gcp "status" "Application status check initiated."
 
+echo "✅ TODO(ricc): to best implement this, we should check the logs for a status of installation to be success or fail."
+
 echo "✅ Foobar app 'status' checked (logged to GCP)."